requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
cryptography==40.0.2
boto3==1.26.0
azure-identity==1.13.0
azure-mgmt-costmanagement==4.0.0
google-cloud-billing==1.9.0
google-cloud-bigquery==3.25.0
google-cloud-compute==1.10.0
PyJWT==2.7.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
whitenoise==6.4.0
anthropic==0.7.0
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
requests==2.31.0
twilio==8.10.0
reportlab==4.0.0
python-dateutil==2.8.2
pytz==2023.3
